requirements.txt
pandas_datareader
pandas
openai 
yfinance
pybind11
cmake
numpy
statsmodels
python-dotenv